corrected a lot of errors with gcc4.3 and peerfuse is compatible with c++0x
git-svn-id: https://piggledy.org/svn/peerfuse/trunk@436 f7b84f85-2e04-4de5-9e99-c4612dc0539c

+++ common/network_base.cpp
@@ -346,8 +346,8 @@ Peer* NetworkBase::Start(MyConfig* conf)
 	/* Listen a TCP port */
 	ConfigSection* section = conf->GetSection("listen");
 
-	Listen(section->GetItem("port")->Integer(),
-		section->GetItem("bind")->String().c_str());
+	Listen(static_cast<uint16_t>(section->GetItem("port")->Integer()),
+	       section->GetItem("bind")->String().c_str());
 
 	/* Connect to other servers */
 	Peer* peer = 0;
@@ -356,7 +356,8 @@ Peer* NetworkBase::Start(MyConfig* conf)
 	{
 		try
 		{
-			peer = Connect((*it)->GetItem("host")->String(), (*it)->GetItem("port")->Integer());
+			peer = Connect((*it)->GetItem("host")->String(),
+			               static_cast<uint16_t>((*it)->GetItem("port")->Integer()));
 
 			/* If Connect() doesn't raise any exception, we are connected and we leave loop */
 			break;
